Show table entry using default port in README
Remove the port specification from one of the example table entries in
the README file and add a comment explaining that the port number from
the listener will be used.

I'm looking to set this up to proxy both http and https traffic using
the same mappings from host name to backend, so I was curious if I'd
need to duplicate that information. Reading the source of the man page I
learned that this feature would let me avoid the duplication, but future
people looking at this may find it nice to be able to learn that from
the README file.